+++ agrest-engine/src/test/java/io/agrest/exp/parser/ExpExistsTest.java
@@ -13,7 +13,8 @@ class ExpExistsTest {
     @ParameterizedTest
     @ValueSource(strings = {
             "exists a",
-            "exists a.b"})
+            "exists a.b",
+    })
     public void parse(String expString) {
         assertEquals(ExpExists.class, Exp.parse(expString).getClass());
     }
@@ -21,14 +22,19 @@ class ExpExistsTest {
     @ParameterizedTest
     @CsvSource(delimiter = '|', value = {
             "exists a|exists a",
-            "exists a.b|exists a.b"
+            "exists a.b|exists a.b",
+            "exists a.b or a.c in (1, 2, 3)|(exists a.b) or (a.c in (1, 2, 3))",
     })
     public void parsedToString(String expString, String expected) {
         assertEquals(expected, Exp.parse(expString).toString());
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"exists", "exists()", "exists (name == 'test')"})
+    @ValueSource(strings = {
+            "exists",
+            "exists()",
+            "exists (name == 'test')",
+    })
     public void parseInvalidGrammar(String expString) {
         assertThrows(AgException.class, () -> Exp.parse(expString));
     }